backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
pymongo
requests
langchain
langchain_groq
